changed damage to float to fix laser damage on object

+++ Employee of the month/Assets/Laser/Scripts/Laser.cs
@@ -128,7 +128,7 @@ public class Laser : MonoBehaviour
         }
         if (collider.gameObject.GetComponent<ItemBreak>() != null)
         {
-            collider.gameObject.GetComponent<ItemBreak>().TakeDamage((int)damage);
+            collider.gameObject.GetComponent<ItemBreak>().TakeDamage(damage);
         }
     }
